Now the script saves the type of visual mode that the user was in when they made the mark, and restores it.
I also needed to change the part that takes the user out of visual mode upon
creating the mark. Typing just v when you are in either line or block visual
only brings it back to regular visual. So no it is escape because that takes it
back to normal mode regardless of which visual mode you were in.